pyproject.toml
[project]
name = "markdown-to-mysql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=46.0.5",
    "fastapi>=0.135.1",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]
